requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.6.0
    # via pydantic
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
idna==3.7
    # via requests
pydantic==2.7.1
    # via nexthink-api (pyproject.toml)
pydantic-core==2.18.2
    # via pydantic
pyyaml==6.0.1
    # via nexthink-api (pyproject.toml)
requests==2.31.0
    # via nexthink-api (pyproject.toml)
typing-extensions==4.11.0
    # via
    #   pydantic
    #   pydantic-core
urllib3==2.2.1
    # via
    #   nexthink-api (pyproject.toml)
    #   requests
